refactor: extract the types of jobs into another dedicated type file
